refactor name of  duckduckgo
