+++ graph/component/__init__.py
@@ -10,7 +10,7 @@ from .message import Message, MessageParam
 from .rewrite import RewriteQuestion, RewriteQuestionParam
 from .keyword import KeywordExtract, KeywordExtractParam
 from .baidu import Baidu, BaiduParam
-from .duckduckgosearch import DuckDuckGoSearch, DuckDuckGoSearchParam
+from .duckduckgo import DuckDuckGo, DuckDuckGoParam
 
 
 def component_class(class_name):
